+++ lib/Slop/Resource.hs
@@ -1,0 +1,45 @@
+module Slop.Resource
+  ( Resource(..)
+  , withResource
+  , releaseResource
+  , Texture
+  , Font
+  , Text
+  , loadTexture
+  , destroyTexture
+  , loadFont
+  , loadFontSDF
+  , closeFont
+  , createText
+  , destroyText
+  , AssetId(..)
+  , AnyAssetId(..)
+  , AssetStatus(..)
+  , AssetUpdate
+  , AssetLoader(..)
+  , AssetThread(..)
+  , TextureAsset(..)
+  , FontAsset(..)
+  , SdfFontAsset(..)
+  , TextAsset(..)
+  , loadAsset
+  , loadAssetResult
+  , loadAssetAsync
+  , awaitAsset
+  , awaitAssetResult
+  , awaitAssetUpdate
+  , awaitAssetUpdateResult
+  , getAsset
+  , getAssetReady
+  , assetReady
+  , getAssetStatus
+  , removeAsset
+  , removeAssets
+  , removeAssets_
+  , removeAllAssets
+  , enableHotReload
+  , disableHotReload
+  , reloadAssetAsync
+  ) where
+
+import Slop.Internal